Refactoring in order to get internal_interpreter.kt smaller
