+++ rpgJavaInterpreter-core/src/main/kotlin/com/smeup/rpgparser/interpreter/movel.kt
@@ -3,10 +3,10 @@ package com.smeup.rpgparser.interpreter
 import com.smeup.rpgparser.parsing.ast.*
 import java.math.BigDecimal
 
-private fun assignStringToString(operationExtender: String?, target: AssignableExpression, valueExpression: Expression, interpreterCoreHelper: InterpreterCoreHelper): Value {
-    var newValue = interpreterCoreHelper.interpret(valueExpression)
+private fun assignStringToString(operationExtender: String?, target: AssignableExpression, valueExpression: Expression, interpreterCore: InterpreterCore): Value {
+    var newValue = interpreterCore.interpret(valueExpression)
     if (valueExpression is AllExpr) {
-        return interpreterCoreHelper.assign(target, newValue)
+        return interpreterCore.assign(target, newValue)
     }
     if (newValue is NumberValue) {
         newValue = newValue.numberToString()
@@ -17,7 +17,7 @@ private fun assignStringToString(operationExtender: String?, target: AssignableE
         newValue = newValue.takeFirst(target.size())
     } else if (valueSize < targetSize) {
         val append = if (operationExtender == null) {
-            val value = interpreterCoreHelper.interpret(target)
+            val value = interpreterCore.interpret(target)
             require(value is StringValue)
             StringValue.padded(value.value, targetSize).takeLast(targetSize - valueSize)
         } else {
@@ -25,14 +25,14 @@ private fun assignStringToString(operationExtender: String?, target: AssignableE
         }
         newValue = newValue.concatenate(append)
     }
-    return interpreterCoreHelper.assign(target, newValue)
+    return interpreterCore.assign(target, newValue)
 }
 
-private fun assignNumberToNumber(operationExtender: String?, target: AssignableExpression, valueExpression: Expression, interpreterCoreHelper: InterpreterCoreHelper): Value {
-    val newValue = interpreterCoreHelper.interpret(valueExpression) as NumberValue
+private fun assignNumberToNumber(operationExtender: String?, target: AssignableExpression, valueExpression: Expression, interpreterCore: InterpreterCore): Value {
+    val newValue = interpreterCore.interpret(valueExpression) as NumberValue
     val targetType = target.type() as NumberType
     val newDecimalValue = DecimalValue(BigDecimal(newValue.bigDecimal.unscaledValue(), targetType.decimalDigits))
-    return interpreterCoreHelper.assign(target, newDecimalValue)
+    return interpreterCore.assign(target, newDecimalValue)
 }
 
 fun size(valueExpression: Expression): Int {
@@ -57,16 +57,16 @@ private fun NumberValue.numberToString(): Value {
     return StringValue(value)
 }
 
-fun movel(operationExtender: String?, target: AssignableExpression, value: Expression, interpreterCoreHelper: InterpreterCoreHelper): Value {
+fun movel(operationExtender: String?, target: AssignableExpression, value: Expression, interpreterCore: InterpreterCore): Value {
     if (value is FigurativeConstantRef) {
-        return interpreterCoreHelper.assign(target, interpreterCoreHelper.interpret(value))
+        return interpreterCore.assign(target, interpreterCore.interpret(value))
     }
     val valueType = value.type()
     if (baseType(target.type()) is StringType && (valueType is StringType || valueType is NumberType || valueType is FigurativeType)) {
-        return assignStringToString(operationExtender, target, value, interpreterCoreHelper)
+        return assignStringToString(operationExtender, target, value, interpreterCore)
     }
     if (baseType(target.type()) is NumberType && (valueType is NumberType || valueType is FigurativeType)) {
-        return assignNumberToNumber(operationExtender, target, value, interpreterCoreHelper)
+        return assignNumberToNumber(operationExtender, target, value, interpreterCore)
     }
     throw IllegalArgumentException("Cannot assign ${valueType::class.qualifiedName} to ${target.type()::class.qualifiedName}")
 }
